// src/data-structures/basic/lru-cache/lru-cache-types.ts
export interface LRUCacheStats {
  /** Total number of cache hits */
  hits: number;
  /** Total number of cache misses */
  misses: number;
  /** Total number of cache sets */
  sets: number;
  /** Total number of cache deletes */
  deletes: number;
  /** Total number of evictions */
  evictions: number;
  /** Current number of items in cache */
  size: number;
  /** Maximum size reached */
  maxSize: number;
  /** Hit rate (hits / (hits + misses)) */
  hitRate: number;
  /** Average access time in milliseconds */
  averageAccessTime: number;
  /** Most frequently accessed key */
  mostAccessedKey?: any;
  /** Least recently used key */
  lruKey?: any;
}
// ...
export interface LRUCachePerformanceMetrics {
  /** Average time for get operations in milliseconds */
  averageGetTime: number;
  /** Average time for set operations in milliseconds */
  averageSetTime: number;
  /** Average time for delete operations in milliseconds */
  averageDeleteTime: number;
  /** Total memory usage estimate in bytes */
  estimatedMemoryUsage: number;
  /** Number of cleanup operations performed */
  cleanupCount: number;
  /** Time spent on cleanup operations in milliseconds */
  totalCleanupTime: number;
}